Fix interaction with gem "haml"
Using a plugin that requires "haml" caused this error:
  ':super' is not an ActiveModel-compatible object that returns a valid partial path .

The problem is that with haml, the additional code for the #render function
never gets called. Fix it by overriding in ActionView::Base instead of
ActionView::Helpers::RenderingHelper. It's the same place that haml uses
to do it's magic.

All tests still passing, tested on Redmine 3.1 and 2.3.4

GitHub #7, #8